fix(rgr): fix output flag of rgr commands, fix outputs, fix test status
Fix output path flag of rgr commands by wrapping the path argument with
a pathlib.Path constructor so that an exception is not raised for string
devision with string which is obviously not supported.
Fix regression outputs by appending a new line for every message popped
out of the messages queue.
Fix test statuses which were not up to date with the latest patch that
was added due to test failures not returning a non-zero error code
(i.e. 0 is returned on both fail and pass - therefore logs must be
parsed to determine result).

Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>
Closes: #35, #42